Rename microservice to api
