+++ cmd/api/update-repo.go
@@ -1,10 +1,8 @@
-package microservice
+package api
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
-	"path/filepath"
 
 	"github.com/dolittle/platform-api/pkg/git"
 	gitStorage "github.com/dolittle/platform-api/pkg/platform/storage/git"
@@ -13,9 +11,9 @@ import (
 	"github.com/spf13/viper"
 )
 
-var gitTestCMD = &cobra.Command{
-	Use:   "git-test",
-	Short: "Test git",
+var updateRepoCMD = &cobra.Command{
+	Use:   "update-repo",
+	Short: "Trigger pull on the git repo",
 	Run: func(cmd *cobra.Command, args []string) {
 		logrus.SetFormatter(&logrus.JSONFormatter{})
 		logrus.SetOutput(os.Stdout)
@@ -29,27 +27,8 @@ var gitTestCMD = &cobra.Command{
 			gitRepoConfig,
 		)
 
-		dir := filepath.Join(gitRepo.Directory, "dev", "453e04a7-4f9d-42f2-b36c-d51fa2c83fa3")
-
-		err := os.MkdirAll(dir, 0755)
+		err := gitRepo.Pull()
 		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		microserviceID := "test"
-		data := []byte(`hi 1`)
-		filename := filepath.Join(dir, fmt.Sprintf("ms_%s.json", microserviceID))
-		err = ioutil.WriteFile(filename, data, 0644)
-		if err != nil {
-			fmt.Println("writeFile")
-			fmt.Println(err)
-			return
-		}
-
-		err = gitRepo.CommitPathAndPush(filename, "upsert microservice")
-		if err != nil {
-			fmt.Println("CommitPathAndPush")
 			fmt.Println(err)
 			return
 		}
