+++ cmd/root.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/dolittle/platform-api/cmd/microservice"
+	"github.com/dolittle/platform-api/cmd/api"
 	"github.com/dolittle/platform-api/cmd/rawdatalog"
 	"github.com/dolittle/platform-api/cmd/tools"
 	"github.com/spf13/cobra"
@@ -21,7 +21,7 @@ func Execute() {
 
 func init() {
 	viper.AutomaticEnv()
-	rootCmd.AddCommand(microservice.RootCmd)
+	rootCmd.AddCommand(api.RootCmd)
 	rootCmd.AddCommand(rawdatalog.RootCmd)
 	rootCmd.AddCommand(tools.RootCmd)
